Add incomplete way to specify blog post ids
We can either strip the filename or provide a specific id (with dir still attached).
The problem with the current implementation is that the blog index also needs the same logic. We can, of course, factor this logic out.
Instead, we'll do this in the loader. This was not done initially because the data of each entry isn't validated yet, over there.